requirements.txt
asgiref==3.4.1
django-cors-headers==3.10.1
gunicorn==20.1.0
Pillow==8.4.0
psycopg2-binary==2.9.2
python-dotenv==0.19.2
sqlparse==0.4.2